fix(react): split i18n provider from light hooks
