+++ packages/react/src/ui/volume-indicator/volume-indicator-root.tsx
@@ -4,7 +4,7 @@ import { createInputIndicatorLabels, VolumeIndicatorCore, VolumeIndicatorDataAtt
 import type { ForwardedRef } from 'react';
 import { forwardRef } from 'react';
 
-import { useTranslator } from '../../i18n';
+import { useTranslator } from '../../i18n/instance';
 import type { UIComponentProps } from '../../utils/types';
 import { renderElement } from '../../utils/use-render';
 import { useInputIndicatorRoot } from '../input-indicators/use-input-indicator-root';
